requirements.txt
annotated-types==0.7.0
attrs==25.3.0
blinker==1.9.0
Brotli==1.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
ConfigArgParse==1.7
filelock==3.12.4
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
gevent==24.11.1
geventhttpclient==2.3.3
greenlet==3.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
locust==2.33.2
MarkupSafe==3.0.2
more-itertools==10.6.0
msgpack==1.1.0
outcome==1.3.0.post0
packaging==24.2
playwright==1.51.0
pluggy==1.5.0
pre==1.3.0
psutil==7.0.0
py==1.11.0
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.1.1
PySocks==1.7.1
pytest==8.3.5
pytest-base-url==2.1.0
pytest-playwright==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pyzmq==26.3.0
qase-api-client==1.1.3
qase-api-v2-client==1.1.2
qase-pytest==6.2.1
qase-python-commons==3.3.1
requests==2.32.3
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
text-unidecode==1.3
toml==0.10.2
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
zope.event==5.0
zope.interface==7.2
